Make tabs more a11y
